Fix data type conversion in add-parsing-number-field.js
- Added explicit type casting auction_number::integer for parsing_number field
- Fixes error: column parsing_number is of type integer but expression is of type character varying
- Now properly converts string auction_number to integer parsing_number

+++ add-parsing-number-field.js
@@ -23,7 +23,7 @@ async function addParsingNumberField() {
         // Обновляем существующие записи: parsing_number = auction_number (временно)
         const updateQuery = `
             UPDATE auction_lots 
-            SET parsing_number = auction_number 
+            SET parsing_number = auction_number::integer 
             WHERE parsing_number IS NULL;
         `;
